requirements.txt
rasa>=3.6.0
rasa-sdk>=3.6.0
pymysql>=1.0.2
SQLAlchemy>=1.4.0,<1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
pandas>=1.3.0